% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{pdf_split}
\alias{pdf_split}
\alias{pdf_length}
\alias{pdf_select}
\alias{pdf_compress}
\title{Split PDF}
\usage{
pdf_split(input, output = NULL)

pdf_length(input)

pdf_select(input, pages = 1, output = NULL)

pdf_compress(input, output = NULL)
}
\arguments{
\item{input}{path to the input pdf file}

\item{output}{base path of the output file(s)}
}
\description{
Split a pdf file in multiple pages.
}
